requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.7.4.post0; python_version >= '3.6'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cycler==0.11.0; python_version >= '3.6'
discord.py==1.7.3
fonttools==4.28.3; python_version >= '3.7'
idna==3.3; python_version >= '3.5'
kiwisolver==1.3.2; python_version >= '3.7'
matplotlib==3.5.0
multidict==5.2.0; python_version >= '3.6'
numpy==1.21.4; python_version < '3.10' and platform_machine != 'aarch64' and platform_machine != 'arm64'
packaging==21.3; python_version >= '3.6'
pandas==1.3.4
pillow==8.4.0; python_version >= '3.6'
pyparsing==3.0.6; python_version >= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.3
setuptools-scm==6.3.2; python_version >= '3.6'
setuptools==59.5.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==1.2.2; python_version >= '3.6'
typing-extensions==4.0.1; python_version >= '3.6'
yarl==1.7.2; python_version >= '3.6'
